test-requirements.txt
coverage>=4.0.3
fixtures>=1.4.0
mock>=1.3.0
testtools>=2.0.0
Flask-WebTest>=0.0.7
python-subunit>=1.2.0
os-testr>=0.6.0
flake8>=2.5.4
